fix type of parameters passed to read and write of memort

+++ src/instructions/ld.rs
@@ -14,7 +14,7 @@ impl LD {
     }
 
     pub fn exec(&self, regs: &mut Reg, memory: &Memory) {
-        regs.Rx[self.dr] = memory.read((regs.PC + self.pcoffset) as usize);
+        regs.Rx[self.dr] = memory.read(regs.PC + self.pcoffset);
         regs.update_flag(regs.Rx[self.dr]);
     }
 }
